# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.ChatMessage:
    properties:
      content:
        type: string
      id:
        type: integer
      receiver_id:
        type: string
      sender_id:
        type: string
      timestamp:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  models.LoginResponse:
    properties:
      token:
        type: string
    type: object
  models.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: contact@dratsisama.com
    name: Dr_AtsiSama
    url: http://dratsisama.com
  description: Documentation de l'API pour la gestion des kermesses scolaires avec
    fonctionnalités de stand, tombola, utilisateurs, etc.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  title: Kermisys API Documentation
  version: "1.0"
paths:
  /forgot-password:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Envoie un lien de réinitialisation à l'utilisateur
      parameters:
      - description: Adresse e-mail
        in: formData
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Mot de passe oublié
      tags:
      - Authentification
...
  /login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Authentifie un utilisateur et retourne un jeton JWT
      parameters:
      - description: Nom d'utilisateur
        in: formData
        name: username
        required: true
        type: string
      - description: Mot de passe
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Connexion
      tags:
      - Authentification
  /logout:
    post:
      description: Déconnexion de l'utilisateur
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
      security:
      - Bearer: []
      summary: Déconnexion
      tags:
      - Authentification
...
  /register:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Crée un nouvel utilisateur
      parameters:
      - description: Nom d'utilisateur
        in: formData
        name: username
        required: true
        type: string
      - description: Mot de passe
        in: formData
        name: password
        required: true
        type: string
      - description: Adresse e-mail
        in: formData
        name: email
        required: true
        type: string
      - description: Rôle de l'utilisateur
        in: formData
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Inscription
      tags:
      - Authentification
  /reset-password:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Permet à un utilisateur de réinitialiser son mot de passe
      parameters:
      - description: Token de réinitialisation
        in: formData
        name: token
        required: true
        type: string
      - description: Nouveau mot de passe
        in: formData
        name: newPassword
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Réinitialiser le mot de passe
      tags:
      - Authentification
...
  /ws/history:
    get:
      description: Récupère l'historique des messages entre deux utilisateurs, avec
        pagination
      parameters:
      - description: ID de l'expéditeur
        in: query
        name: sender_id
        required: true
        type: string
      - description: ID du destinataire
        in: query
        name: receiver_id
        required: true
        type: string
      - default: 50
        description: Nombre de messages à récupérer
        in: query
        name: limit
        type: integer
      - default: 1
        description: Numéro de page
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ChatMessage'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Récupérer l'historique des messages
      tags:
      - Chat
  /ws/private:
    get:
      description: Établit une connexion WebSocket pour la discussion privée en temps
        réel entre un organisateur et un gérant de stand
      parameters:
      - description: 'Identifiant de l''utilisateur (ex : username)'
        in: query
        name: userID
        required: true
        type: string
      responses:
        "101":
          description: Switching Protocols" - WebSocket handshake success
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: WebSocket Chat Privé
      tags:
      - WebSocket
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"